refactor: split up commands and add documentation
