+++ packages/integration-tests/README.md
@@ -3,11 +3,6 @@
 Add tests to integration-karma instead of integration-tests unless the tests are related to
 keyboard navigation.
 
-Pro tip: `browser.keys(['Shift', 'Tab'])` keeps you in a state where the shift key is toggled so
-you may run into unexpected results unless you explicitly untoggle the shift key. The
-recommendation is to use the command `browser.keys(['Shift', 'Tab', 'Shift'])` which toggles and
-untoggles the shift key in a single command.
-
 End-to-end tests run in the browser across different modes with the option to run tests remotely on Sauce Labs.
 
 ## Commands
@@ -149,3 +144,9 @@ lwc-integration/src/components/<functional-area>
 ```
 
 Note that `<specific-repro>.spec.js` should always be at the top level under `test-<specific-repro>`.
+
+### Pro tips
+
+-   `browser.keys(['Shift', 'Tab'])` keeps you in a state where the shift key is toggled so you may run into unexpected results unless you explicitly untoggle the shift key. The recommendation is to use the command `browser.keys(['Shift', 'Tab', 'Shift'])` which toggles and untoggles the shift key in a single command.
+
+-   On top of the [commands](https://webdriver.io/docs/api.html) already provided by WebdriverIO, this project also provides [custom command](scripts/commands) for common operation like shadow tree query selection and focus.
